+++ saga-gis/src/tools/io/io_gdal/gdal_import_vrt.cpp
@@ -372,7 +372,7 @@ bool CGDAL_Import_VRT::On_Execute(void)
 
 		for(size_t i=0; i<pGrids.Get_Size(); i++)
 		{
-			pCollection->Add_Grid(i, (CSG_Grid *)pGrids.Get(i), true);
+			pCollection->Add_Grid((double)i, (CSG_Grid *)pGrids.Get(i), true);
 		}
 
 		pList->Add_Item(pCollection);
